+++ app.go
@@ -1,12 +1,31 @@
 package main
 
 import (
-	"tat_gogogo/server"
 	"github.com/gin-gonic/gin"
+	"tat_gogogo/configs"
+	"tat_gogogo/interface/routes"
+	"log"
 )
 
-var HttpServer *gin.Engine
+func run() {
+	var httpServer *gin.Engine
+
+	configuration, err := configs.New()
+
+	if err != nil {
+		log.Panicln("Configuration err", err)
+	}
+
+	httpServer = gin.Default()
+
+	routes.RegisterRoutes(httpServer)
+
+	serverAddr := configuration.Constants.Host + ":" + configuration.Constants.Port
+
+	// listen and serve on 0.0.0.0:8080
+	httpServer.Run(serverAddr)
+}
 
 func main() {
-	server.Run(HttpServer)
+	run()
 }
