+++ domain/model/Curriculum.go
