+++ domain/model/Result.go
