+++ interface/routes/router.go
@@ -1,8 +1,8 @@
 package routes
 
 import (
-	"tat_gogogo/controllers/curriculum"
-	"tat_gogogo/controllers/login"
+	"tat_gogogo/interface/controllers/curriculum"
+	"tat_gogogo/interface/controllers/login"
 
 	"github.com/gin-gonic/gin"
 )
